requirements.txt
arxiv==2.3.1
chromadb==1.3.7
colorlog==6.10.1
huggingface-hub==0.36.0
langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.7
langchain-google-genai==4.1.1
langchain-groq==1.1.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langsmith==0.6.2
pandas==2.3.3
pillow==12.1.0
pypdf==6.4.2
PyMuPDF==1.26.7
pysqlite3-binary==0.5.2
python-dotenv==1.2.1
pytest==9.0.2
replicate==1.0.7
requests==2.32.5
sentence-transformers==5.2.0
streamlit==1.52.2
